large refactor + opensearch

+++ query.sql
@@ -1,17 +1,17 @@
--- name: CreateCategory :one
-INSERT INTO categories (name) VALUES ($1) RETURNING *;
+-- name: CreateTag :one
+INSERT INTO tags (name) VALUES ($1) RETURNING *;
 
--- name: GetCategoryById :one
-SELECT * FROM categories WHERE id = $1;
+-- name: GetTagById :one
+SELECT * FROM tags WHERE id = $1;
 
--- name: GetAllCategories :many
-SELECT * FROM categories;
+-- name: GetTags :many
+SELECT * FROM tags;
 
--- name: UpdateCategory :one
-UPDATE categories SET name = $1 WHERE id = $2 RETURNING *;
+-- name: UpdateTag :one
+UPDATE tags SET name = $1 WHERE id = $2 RETURNING *;
 
--- name: DeleteCategory :exec
-DELETE FROM categories WHERE id = $1;
+-- name: DeleteTag :exec
+DELETE FROM tags WHERE id = $1;
 
 -- name: CreateUser :one
 INSERT INTO users (username, email, password_hash, is_verified, verification_token) VALUES ($1, $2, $3, $4, $5) RETURNING *;
@@ -31,31 +31,35 @@ SELECT user_id,username,email,is_verified FROM users;
 -- name: UpdateUser :one
 UPDATE users SET username = $1, email = $2, password_hash = $3, is_verified = $4, verification_token = $5 WHERE user_id = $6 RETURNING *;
 
--- name: DeleteUser :exec
-DELETE FROM users WHERE user_id = $1;
+-- name: DeleteUser :one
+DELETE FROM users WHERE user_id = $1 RETURNING user_id;
 
 -- name: VerifyUser :exec
 UPDATE users SET is_verified = true, verification_token = null , updated_at = now() WHERE verification_token = $1  RETURNING *;
 
 
 -- name: CreateArticle :one
-INSERT INTO articles (title, content, category_id, user_id) VALUES ($1, $2, $3, $4) RETURNING article_id;
+INSERT INTO articles (title, content, tag_id, user_id) VALUES ($1, $2, $3, $4) RETURNING article_id;
 
 -- name: PublishArticle :exec
 UPDATE articles SET is_published = true, updated_at = now() WHERE article_id = $1;
 
 -- name: GetAllArticles :many
-SELECT a.article_id, a.title, a.content, a.user_id, c.name as category_name, a.created_at, a.updated_at, a.is_published
+SELECT a.article_id, a.title, a.content, a.user_id, t.name as tag_name, a.created_at, a.updated_at, a.is_published
 FROM articles a
-LEFT JOIN categories c ON a.category_id = c.id
-WHERE a.is_published = true;
+LEFT JOIN tags t ON a.tag_id = t.id
+WHERE a.is_published = true
+ORDER BY a.created_at DESC
+LIMIT 10 OFFSET $1;
 
 
 -- name: GetAllArticleByUser :many
-SELECT a.article_id, a.title, a.content, a.user_id, c.name as category_name, a.created_at, a.updated_at, a.is_published
+SELECT a.article_id, a.title, a.content, a.user_id, t.name as Tag_name, a.created_at, a.updated_at, a.is_published
 FROM articles a
-LEFT JOIN categories c ON a.category_id = c.id
-WHERE a.user_id = $1;
+LEFT JOIN Tags t ON a.tag_id = t.id
+WHERE a.user_id = $1
+ORDER BY a.created_at DESC
+LIMIT 10 OFFSET $2;
 
 -- name: GetUserIdByArticleId :one
 SELECT user_id FROM articles WHERE article_id = $1;
